hammer: Use x86-64 indexers

+++ hammer/dispmanager.cpp
@@ -157,7 +157,7 @@ public: // functions
 	// selection list functions
 	intp SelectCount( void );
 	void SelectClear( void );
-	CMapDisp *GetFromSelect( int iSelectList );
+	CMapDisp *GetFromSelect( intp iSelectList );
 
 	void AddToSelect( EditDispHandle_t handle );
 	void RemoveFromSelect( EditDispHandle_t handle );
@@ -487,7 +487,7 @@ void CWorldEditDispMgr::SelectClear( void )
 //-----------------------------------------------------------------------------
 // Purpose:
 //-----------------------------------------------------------------------------
-CMapDisp *CWorldEditDispMgr::GetFromSelect( int iSelectList )
+CMapDisp *CWorldEditDispMgr::GetFromSelect( intp iSelectList )
 {
 	// no assert because the .Element( ) takes care of that!
 	EditDispHandle_t handle = m_SelectList.Element( iSelectList );
